Some progress, need to figure out order of setting pystats deptha and initializing memory

+++ Include/internal/pycore_pystats.h
@@ -12,7 +12,7 @@ extern "C" {
 extern void _Py_StatsOn(void);
 extern void _Py_StatsOff(void);
 extern void _Py_StatsClear(void);
-void _Py_Stats_Maybe_Set_Depth(uint64_t depth);
+void _Py_Stats_Maybe_Set_Depth(int depth);
 extern int _Py_PrintSpecializationStats(int to_file);
 #endif
